requirements.txt
pytest==7.4.4
selenium==4.11.2
python-dotenv==0.21.1
allure-pytest==2.13.5
allure-python-commons==2.13.5

